Bug Fixes / Error Checking for IDs that already exist
Some Error checking to complete the script. This will allow no duplicate IDs of BOTH secondaryTurtleIDs AND secondaryImageIDs.

+ Updated scraping filenames to be more clear of purpose
+ Added duplicate checking for both secondaryIDs in their functions
+ If it doesnt exist yet, insert. If it does then do not insert the duplicate.
+ Instead of reading the file again a second time to insert ImageIDs we appended them into a tuple the first time around. We will insert values from the tuple into the table.

Note: Two package.json were updated in this commit not by me. This was done AUTOMATICALLY. This is to log the change. Not sure if it is important or if it will impact other developers.