Update (CSS)
Major style changes to the landing page and navigation bar. These changes include:

+ Background gradient following FIGMA style
+ Navigation bar with logo, project name and a burger menu (with dropdown style)
+ Significant style changes to form with headers, sub-headers, icons, animations, buttons, borders and placeholder texts for UX
+ Added Title and subheading for landing page prior to form (UX)
+ Drop down menus lead to other pages (YET TO BE RENAMED)

Notes:
There are multiple minor changes still to be tweaked such as drop down menu names and styling. The footer can also be reworked to appear 'better-looking' in another update.

Any criticism appreciated. This is not reflective of the finished product ***