update (/results CSS)
Significantly styled results page to near completion. There is room for plenty more CSS if needed, basics are completed.

These changes include:
+ style.css slight name changes to classes
+ Improved comment input by replacing with 'textarea' to allow multiple lines of input.
+ fixed loading spinner working as intended (1.5second animation)
+ added header, body, #IF for loading spinner and lists to structure results form
+ added new fields "Orientation", "Comment"

- removed map on /details temporarily

Notes:
* Need to nest <style> in .css file (will do that later)
* Multiple fields are null because there are no mock data as of current
* Optional: Need help shortening longitude and latitude to three decimal places
* Optional: Need to see if we can grab location name from maps and fill the extra field i added
* Buttons may need better CSS
* IMPORTANT: MAP NEEDS TO LOAD BEFORE BUTTON PRESS (need additional help)

Pretty sure someone else is working on the images in the blob, was working as of an hour ago but not right now (nothing has changed), so it may or may not display image.