Bug Fixes + new feature
fixed when attempting to upload image, where it would ask for an image to be uploaded even if one is already selected.

Added feature so images can be dragged and dropped onto the web app to be uploaded.

Fixed an issue when the results page first loaded, the map was blank until you pressed the next turtle button.

Fixed an issue where if the user refreshed the results page it showed an internal server error.